Problem: code is not idomatic
Five-phase refactoring aligned with AGENTS.md design principles:

Phase 1 — Shared math module:
- New pydefi/_math.py: apply_slippage, slippage_to_*, price_impact
- Removes 3 duplicate _apply_slippage/_slippage_to_* from base classes

Phase 2 — Remove abstract base classes:
- Delete amm/base.py, bridge/base.py (no more ABCs in the project)
- Inline super().__init__ into 20+ concrete classes (UniswapV2/V3, Curve,
Raydium, Across, CCIP, CCTP, Eureka, GasZip, LayerZeroOFT, LucidBridge,
Mayan, Relay, Stargate, OneInch, OKX, ParaSwap, ZeroX, OpenOcean,
UniswapAPI, Jupiter, JupiterSwapV2, GeckoTerminal, Subgraph, etc.)
- Convert BasePool(ABC) → BasePool (plain class)

Phase 3 — Move RouteDAG out of types.py:
- New pydefi/pathfinder/dag.py (RouteDAG + _freeze_actions + builder helpers)
- types.py now holds pure data models only (~200 lines removed)
- MAX_BPS deduplicated; all importers use _math.MAX_BPS

Phase 4 — Lift staticmethods to module-level pure functions:
- v2_get_amount_out, v2_get_amount_in, v2_spot_price (UniswapV2)
- v3_sqrt_price_to_price, v3_encode_path (UniswapV3)
- _dag_leg_weights, _edges_to_dag, _estimate_price_impact (Router)
- _parse_price_impact (OpenOcean), _encode_transfer_message (LucidBridge)
- build_graph (pool_data) elevated to module-level

Bonus:
- AGENTS.md: 7 design principles + imports-at-top rule
- Fix lazy imports: eureka.py (time), indexer.py (ERC20)
- All 789 non-live/non-fork tests pass with zero regressions
