+++ pydefi/aggregator/__init__.py
@@ -1,6 +1,6 @@
 """DEX aggregator API integrations."""
 
-from pydefi.aggregator.base import AggregatorQuote, BaseAggregator
+from pydefi.aggregator.base import AggregatorQuote
 from pydefi.aggregator.jupiter import Jupiter, JupiterSwapV2
 from pydefi.aggregator.okx import OKX
 from pydefi.aggregator.oneinch import OneInch
@@ -10,7 +10,6 @@ from pydefi.aggregator.uniswap import UniswapAPI
 from pydefi.aggregator.zerox import ZeroX
 
 __all__ = [
-    "BaseAggregator",
     "AggregatorQuote",
     "Jupiter",
     "JupiterSwapV2",
