+++ pydefi/aggregator/oneinch.py
@@ -11,12 +11,13 @@ from typing import Any
 
 import aiohttp
 
-from pydefi.aggregator.base import AggregatorQuote, BaseAggregator
+from pydefi._math import slippage_to_fraction, slippage_to_percent
+from pydefi.aggregator.base import AggregatorQuote
 from pydefi.exceptions import AggregatorError
 from pydefi.types import SwapRoute, SwapStep, Token, TokenAmount
 
 
-class OneInch(BaseAggregator):
+class OneInch:
     """1inch Aggregation Protocol API client.
 
     Args:
@@ -33,7 +34,8 @@ class OneInch(BaseAggregator):
         api_key: str | None = None,
         base_url: str | None = None,
     ) -> None:
-        super().__init__(chain_id, api_key)
+        self.chain_id = chain_id
+        self.api_key = api_key
         self._base_url = base_url or self._DEFAULT_BASE_URL
 
     @property
@@ -133,7 +135,7 @@ class OneInch(BaseAggregator):
             "dst": token_out.encoded_address,
             "amount": str(amount_in.amount),
             "from": from_address,
-            "slippage": self._slippage_to_percent(slippage_bps),
+            "slippage": slippage_to_percent(slippage_bps),
             **kwargs,
         }
         data = await self._get("swap", params)
