+++ pydefi/aggregator/paraswap.py
@@ -11,12 +11,12 @@ from typing import Any
 
 import aiohttp
 
-from pydefi.aggregator.base import AggregatorQuote, BaseAggregator
+from pydefi.aggregator.base import AggregatorQuote
 from pydefi.exceptions import AggregatorError
 from pydefi.types import SwapRoute, SwapStep, Token, TokenAmount
 
 
-class ParaSwap(BaseAggregator):
+class ParaSwap:
     """ParaSwap DEX aggregator API client.
 
     Args:
@@ -33,7 +33,8 @@ class ParaSwap(BaseAggregator):
         api_key: str | None = None,
         base_url: str | None = None,
     ) -> None:
-        super().__init__(chain_id, api_key)
+        self.chain_id = chain_id
+        self.api_key = api_key
         self._base_url = base_url or self._DEFAULT_BASE_URL
 
     @property
