+++ pydefi/aggregator/zerox.py
@@ -11,7 +11,8 @@ from typing import Any
 
 import aiohttp
 
-from pydefi.aggregator.base import AggregatorQuote, BaseAggregator
+from pydefi._math import slippage_to_fraction, slippage_to_percent
+from pydefi.aggregator.base import AggregatorQuote
 from pydefi.exceptions import AggregatorError
 from pydefi.types import Address, SwapRoute, SwapStep, Token, TokenAmount
 
@@ -27,7 +28,7 @@ _CHAIN_URLS: dict[int, str] = {
 }
 
 
-class ZeroX(BaseAggregator):
+class ZeroX:
     """0x Protocol DEX aggregator API client.
 
     The 0x Swap API provides smart order routing across liquidity sources.
@@ -44,7 +45,8 @@ class ZeroX(BaseAggregator):
         api_key: str | None = None,
         base_url: str | None = None,
     ) -> None:
-        super().__init__(chain_id, api_key)
+        self.chain_id = chain_id
+        self.api_key = api_key
         self._base_url = base_url or _CHAIN_URLS.get(chain_id, "https://api.0x.org")
 
     @property
@@ -116,7 +118,7 @@ class ZeroX(BaseAggregator):
             "sellToken": amount_in.token.address,
             "buyToken": token_out.address,
             "sellAmount": str(amount_in.amount),
-            "slippagePercentage": self._slippage_to_fraction(slippage_bps),
+            "slippagePercentage": slippage_to_fraction(slippage_bps),
             **kwargs,
         }
         data = await self._get("swap/v1/quote", params)
