+++ pydefi/amm/__init__.py
@@ -1,6 +1,5 @@
 """AMM (Automated Market Maker) integrations."""
 
-from pydefi.amm.base import BaseAMM, BaseSolanaAMM
 from pydefi.amm.curve import CurvePool
 from pydefi.amm.raydium import Raydium
 from pydefi.amm.uniswap_v2 import UniswapV2
@@ -15,8 +14,6 @@ from pydefi.amm.universal_router import (
 )
 
 __all__ = [
-    "BaseAMM",
-    "BaseSolanaAMM",
     "UniswapV2",
     "UniswapV3",
     "CurvePool",
