+++ pydefi/bridge/across.py
@@ -22,15 +22,15 @@ from typing import Any
 import aiohttp
 from web3 import AsyncWeb3
 
+from pydefi._math import apply_slippage
 from pydefi.abi.bridge import ACROSS_SPOKE_POOL
-from pydefi.bridge.base import BaseBridge
 from pydefi.exceptions import BridgeError
 from pydefi.types import ZERO_ADDRESS, Address, BridgeQuote, Token, TokenAmount
 
 _ACROSS_API_BASE = "https://app.across.to/api"
 
 
-class Across(BaseBridge):
+class Across:
     """Across Protocol cross-chain bridge integration.
 
     Args:
@@ -50,7 +50,8 @@ class Across(BaseBridge):
         spoke_pool_address: str,
         api_base_url: str = _ACROSS_API_BASE,
     ) -> None:
-        super().__init__(src_chain_id, dst_chain_id)
+        self.src_chain_id = src_chain_id
+        self.dst_chain_id = dst_chain_id
         self.w3 = w3
         self.spoke_pool_address = spoke_pool_address
         self._api_base = api_base_url.rstrip("/")
@@ -158,7 +159,7 @@ class Across(BaseBridge):
 
         total_relay_fee_pct = int(fees_data.get("totalRelayFee", {}).get("pct", "0"))
         fee_raw = amount_in.amount * total_relay_fee_pct // (10**18)
-        output_amount = self._apply_slippage(max(0, amount_in.amount - fee_raw), slippage_bps)
+        output_amount = apply_slippage(max(0, amount_in.amount - fee_raw), slippage_bps)
         quote_timestamp = int(fees_data.get("timestamp", 0))
         fill_deadline = quote_timestamp + 18_000  # 5 hours
