+++ pydefi/pathfinder/router.py
@@ -18,7 +18,54 @@ from decimal import Decimal
 
 from pydefi.exceptions import NoRouteFoundError
 from pydefi.pathfinder.graph import PoolEdge, PoolGraph
-from pydefi.types import MAX_BPS, Address, RouteDAG, RouteSplit, RouteSwap, SwapRoute, SwapStep, Token, TokenAmount
+from pydefi._math import MAX_BPS
+from pydefi.pathfinder.dag import RouteDAG
+from pydefi.types import Address, RouteSplit, RouteSwap, SwapRoute, SwapStep, Token, TokenAmount
+
+
+# ---------------------------------------------------------------------------
+# Module-level helpers
+# ---------------------------------------------------------------------------
+
+
+def _dag_leg_weights(dag: RouteDAG) -> list[int]:
+    """Return ``fraction_bps`` for each split leg, or ``[10000]`` for a linear DAG."""
+    payload = dag.to_dict()
+    if payload["actions"] and isinstance(payload["actions"][0], RouteSplit):
+        return [leg.fraction_bps for leg in payload["actions"][0].legs]
+    return [MAX_BPS]
+
+
+def _edges_to_dag(token_in: Token, edges: list["PoolEdge"]) -> RouteDAG:
+    dag = RouteDAG().from_token(token_in)
+    for edge in edges:
+        dag.swap(edge.token_out, edge)
+    return dag
+
+
+def _estimate_price_impact(edges: list["PoolEdge"], amount_in: int) -> Decimal:
+    """Estimate cumulative price impact across a multi-hop path.
+
+    Delegates per-hop impact estimation to each edge's polymorphic
+    :meth:`~pydefi.pathfinder.graph.PoolEdge.estimate_price_impact` method.
+
+    If a hop returns ``Decimal('NaN')`` (impact unestimable) and no other
+    hop yields a positive estimate, the cumulative result is
+    ``Decimal('NaN')`` to signal "impact unknown" rather than "zero impact".
+    """
+    total_impact = Decimal(0)
+    current_amount = amount_in
+    has_unestimated_hop = False
+    for edge in edges:
+        hop_impact = edge.estimate_price_impact(current_amount)
+        if hop_impact.is_nan():
+            has_unestimated_hop = True
+        else:
+            total_impact += hop_impact
+        current_amount = edge.amount_out(current_amount)
+    if has_unestimated_hop and total_impact == Decimal(0):
+        return Decimal("NaN")
+    return min(total_impact, Decimal(1))
 
 
 class Router:
@@ -176,8 +223,8 @@ class Router:
             steps=steps,
             amount_in=amount_in,
             amount_out=TokenAmount(token=token_out, amount=final_amount),
-            price_impact=self._estimate_price_impact(final_path, amount_in.amount),
-            dag=self._edges_to_dag(src, final_path),
+            price_impact=_estimate_price_impact(final_path, amount_in.amount),
+            dag=_edges_to_dag(src, final_path),
         )
 
     def _find_best_route_gas_aware(
@@ -226,8 +273,8 @@ class Router:
             ],
             amount_in=amount_in,
             amount_out=TokenAmount(token=token_out, amount=final_amount),
-            price_impact=self._estimate_price_impact(path, amount_in.amount),
-            dag=self._edges_to_dag(amount_in.token, path),
+            price_impact=_estimate_price_impact(path, amount_in.amount),
+            dag=_edges_to_dag(amount_in.token, path),
         )
 
     def find_all_routes(
@@ -308,8 +355,8 @@ class Router:
                         steps=steps,
                         amount_in=amount_in,
                         amount_out=TokenAmount(token=token_out, amount=current_amount),
-                        price_impact=self._estimate_price_impact(path, amount_in.amount),
-                        dag=self._edges_to_dag(src, path),
+                        price_impact=_estimate_price_impact(path, amount_in.amount),
+                        dag=_edges_to_dag(src, path),
                     )
                 )
                 return
@@ -433,7 +480,7 @@ class Router:
 
         if len(legs) == 1:
             _, edges = legs[0]
-            return self._edges_to_dag(amount_in.token, edges)
+            return _edges_to_dag(amount_in.token, edges)
 
         dag = RouteDAG().from_token(amount_in.token)
         dag.split()
@@ -546,7 +593,7 @@ class Router:
                     steps=steps,
                     amount_in=amount_in,
                     amount_out=TokenAmount(token=token_out, amount=final_amount),
-                    price_impact=self._estimate_price_impact(final_path, amount_in.amount),
+                    price_impact=_estimate_price_impact(final_path, amount_in.amount),
                 )
             )
         return routes
@@ -654,61 +701,3 @@ class Router:
                     total += self._simulate_actions(leg.actions, leg_amount)
                 amount = total
         return amount
-
-    @staticmethod
-    def dag_leg_weights(dag: RouteDAG) -> list[int]:
-        """Return ``fraction_bps`` for each split leg, or ``[10000]`` for a linear DAG."""
-
-        payload = dag.to_dict()
-        if payload["actions"] and isinstance(payload["actions"][0], RouteSplit):
-            return [leg.fraction_bps for leg in payload["actions"][0].legs]
-        return [MAX_BPS]
-
-    @staticmethod
-    def _edges_to_dag(token_in: Token, edges: list[PoolEdge]) -> RouteDAG:
-        dag = RouteDAG().from_token(token_in)
-        for edge in edges:
-            dag.swap(edge.token_out, edge)
-        return dag
-
-    @staticmethod
-    def _estimate_price_impact(edges: list[PoolEdge], amount_in: int) -> Decimal:
-        """Estimate cumulative price impact across a multi-hop path.
-
-        Delegates per-hop impact estimation to each edge's polymorphic
-        :meth:`~pydefi.pathfinder.graph.PoolEdge.estimate_price_impact` method,
-        allowing each pool type (V2, V3, Curve, …) to implement its own model:
-
-        * :class:`~pydefi.pathfinder.graph.PoolEdge` (V2-style): uses
-          ``amount_in / (reserve_in + amount_in)``.
-        * :class:`~pydefi.pathfinder.graph.V3PoolEdge`: uses virtual reserves
-          derived from ``sqrtPriceX96`` / ``liquidity``.
-
-        If a hop returns ``Decimal('NaN')`` (impact unestimable) and no other
-        hop yields a positive estimate, the cumulative result is
-        ``Decimal('NaN')`` to signal "impact unknown" rather than "zero impact".
-
-        Args:
-            edges: Ordered pool edges in the route.
-            amount_in: Input amount at the first hop.
-
-        Returns:
-            Estimated cumulative price impact in ``[0, 1]``, or
-            ``Decimal('NaN')`` if the entire path consists of pools where
-            impact cannot be estimated.
-        """
-        total_impact = Decimal(0)
-        current_amount = amount_in
-        has_unestimated_hop = False
-        for edge in edges:
-            hop_impact = edge.estimate_price_impact(current_amount)
-            if hop_impact.is_nan():
-                has_unestimated_hop = True
-            else:
-                total_impact += hop_impact
-            # Always propagate the simulated amount forward so that later hops
-            # use the correct intermediate amount.
-            current_amount = edge.amount_out(current_amount)
-        if has_unestimated_hop and total_impact == Decimal(0):
-            return Decimal("NaN")
-        return min(total_impact, Decimal(1))
