+++ pydefi/pool_data/__init__.py
@@ -1,12 +1,12 @@
 """Pool data provider integrations (GeckoTerminal, subgraph)."""
 
-from pydefi.pool_data.base import BasePoolDataProvider, PoolData
+from pydefi.pool_data.base import PoolData, build_graph
 from pydefi.pool_data.geckoterminal import GeckoTerminal
 from pydefi.pool_data.subgraph import Subgraph, UniswapV2Subgraph, UniswapV3Subgraph
 
 __all__ = [
-    "BasePoolDataProvider",
     "PoolData",
+    "build_graph",
     "GeckoTerminal",
     "Subgraph",
     "UniswapV2Subgraph",
