+++ pydefi/pool_data/base.py
@@ -1,10 +1,9 @@
 """
-Base class and data types for pool data provider integrations.
+Data types for pool data provider integrations.
 """
 
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
 from dataclasses import dataclass, field
 
 from pydefi.pathfinder.graph import PoolEdge, PoolGraph, V3PoolEdge
@@ -115,104 +114,14 @@ class PoolData:
         return [edge_0_to_1, edge_1_to_0]
 
 
-class BasePoolDataProvider(ABC):
-    """Abstract base class for pool data providers.
+def build_graph(pools: list[PoolData]) -> PoolGraph:
+    """Convert a list of :class:`PoolData` objects into a :class:`~pydefi.pathfinder.graph.PoolGraph`.
 
-    Sub-classes implement fetching pool state from external sources (e.g. the
-    GeckoTerminal REST API or a protocol-specific subgraph) and expose a
-    uniform interface that the rest of pydefi can consume.
-
-    The :meth:`build_graph` helper converts a list of :class:`PoolData`
-    objects into a :class:`~pydefi.pathfinder.graph.PoolGraph` ready for
-    use with :class:`~pydefi.pathfinder.router.Router`.
+    Each :class:`PoolData` entry contributes two directed edges (one per
+    swap direction) to the graph via :meth:`PoolData.to_pool_edges`.
     """
-
-    @property
-    @abstractmethod
-    def provider_name(self) -> str:
-        """Human-readable name of this data provider."""
-
-    @abstractmethod
-    async def get_pool(self, pool_address: str) -> PoolData:
-        """Fetch data for a single pool by its on-chain address.
-
-        Args:
-            pool_address: Checksummed (or lowercase) pool contract address.
-
-        Returns:
-            A populated :class:`PoolData` object.
-
-        Raises:
-            :class:`~pydefi.exceptions.PoolDataError`: On API errors or when
-                the pool is not found.
-        """
-
-    @abstractmethod
-    async def get_top_pools(self, limit: int = 100) -> list[PoolData]:
-        """Fetch the top pools ranked by liquidity / TVL.
-
-        Args:
-            limit: Maximum number of pools to return.
-
-        Returns:
-            A list of :class:`PoolData` objects.
-        """
-
-    @abstractmethod
-    async def get_pools_for_token(self, token_address: str, limit: int = 100) -> list[PoolData]:
-        """Fetch pools that contain a specific token.
-
-        Args:
-            token_address: ERC-20 token address to filter by.
-            limit: Maximum number of pools to return.
-
-        Returns:
-            A list of :class:`PoolData` objects.
-        """
-
-    async def get_pools_for_tokens(self, token_addresses: list[str], limit: int = 100) -> list[PoolData]:
-        """Fetch pools that contain any of the given tokens.
-
-        Default implementation fans out to :meth:`get_pools_for_token` for
-        each address and deduplicates results by pool address.  Providers
-        that offer a native batch endpoint (e.g. GeckoTerminal) should
-        override this method for better efficiency.
-
-        Args:
-            token_addresses: List of ERC-20 token addresses to query.
-            limit: Maximum total number of pools to return.
-
-        Returns:
-            A deduplicated list of up to *limit* :class:`PoolData` objects.
-        """
-        pools: list[PoolData] = []
-        seen: set[str] = set()
-        for address in token_addresses:
-            if len(pools) >= limit:
-                break
-            for pool in await self.get_pools_for_token(address, limit=limit):
-                if pool.pool_address not in seen:
-                    seen.add(pool.pool_address)
-                    pools.append(pool)
-                    if len(pools) >= limit:
-                        break
-        return pools
-
-    def build_graph(self, pools: list[PoolData]) -> PoolGraph:
-        """Convert a list of :class:`PoolData` objects into a :class:`~pydefi.pathfinder.graph.PoolGraph`.
-
-        Each :class:`PoolData` entry contributes two directed edges (one per
-        swap direction) to the graph via :meth:`PoolData.to_pool_edges`.
-
-        Args:
-            pools: Pool data objects to add to the graph.
-
-        Returns:
-            A :class:`~pydefi.pathfinder.graph.PoolGraph` populated with all
-            provided pools.
-        """
-        graph = PoolGraph()
-        for pool in pools:
-            for edge in pool.to_pool_edges():
-                graph.add_pool(edge)
-        return graph
+    graph = PoolGraph()
+    for pool in pools:
+        for edge in pool.to_pool_edges():
+            graph.add_pool(edge)
+    return graph
