+++ tests/live/test_okx_router_encoder_fork.py
@@ -33,7 +33,7 @@ from pydefi.aggregator.okx_router_encoder import (
     route_dag_to_router_paths,
 )
 from pydefi.pathfinder.graph import PoolEdge
-from pydefi.types import RouteDAG
+from pydefi.pathfinder.dag import RouteDAG
 from tests.addrs import (
     DAI,
     PAIR_USDC_DAI,
