+++ tests/live/test_router_live.py
@@ -12,7 +12,7 @@ from eth_contract import Contract
 from pydefi.exceptions import NoRouteFoundError
 from pydefi.pathfinder.graph import PoolGraph, V3PoolEdge
 from pydefi.pathfinder.router import Router
-from pydefi.types import RouteDAG, RouteSplit, TokenAmount
+from pydefi.types import RouteSplit, TokenAmount
 from pydefi.vm import build_quote_program_for_dag
 from tests.addrs import (
     DAI,
