+++ tests/test_pathfinder.py
@@ -7,8 +7,9 @@ import pytest
 
 from pydefi.exceptions import NoRouteFoundError
 from pydefi.pathfinder.graph import PoolEdge, PoolGraph, V3PoolEdge
-from pydefi.pathfinder.router import Router
-from pydefi.types import Address, ChainId, RouteDAG, RouteSwap, Token, TokenAmount
+from pydefi.pathfinder.router import Router, _estimate_price_impact
+from pydefi.types import Address, ChainId, RouteSwap, Token, TokenAmount
+from pydefi.pathfinder.dag import RouteDAG
 from tests.addrs import DAI, USDC, WETH
 
 # ---------------------------------------------------------------------------
@@ -549,7 +550,7 @@ class TestRouter:
     def test_price_impact_zero_reserves(self):
         """Routes through zero-reserve pools (e.g. V3) should return NaN (unestimated)."""
         edges = [PoolEdge(WETH, USDC, POOL_A, "UniswapV2", reserve_in=0, reserve_out=0)]
-        impact = Router._estimate_price_impact(edges, 10**18)
+        impact = _estimate_price_impact(edges, 10**18)
         assert impact.is_nan()
 
     def test_price_impact_nonzero(self):
@@ -564,7 +565,7 @@ class TestRouter:
                 fee_bps=30,
             )
         ]
-        impact = Router._estimate_price_impact(edges, 10**18)
+        impact = _estimate_price_impact(edges, 10**18)
         assert Decimal(0) < impact < Decimal(1)
 
     def test_find_best_route_with_gas_prefers_lower_hop_when_gas_is_high(self):
