+++ tests/test_program.py
@@ -7,7 +7,8 @@ from eth_contract.contract import ContractFunction
 from hexbytes import HexBytes
 
 from pydefi.bridge.eureka import encode_send_transfer_calldata
-from pydefi.types import Address, BasePool, ChainId, RouteBridge, RouteDAG, SwapProtocol, Token
+from pydefi.types import Address, BasePool, ChainId, RouteBridge, SwapProtocol, Token
+from pydefi.pathfinder.dag import RouteDAG
 from pydefi.vm import (
     IIBC_SENDER_CALLBACKS_INTERFACE_ID,
     Operand,
