+++ tests/test_solana.py
@@ -5,10 +5,10 @@ from unittest.mock import AsyncMock, patch
 
 import pytest
 
+from pydefi._math import apply_slippage
 from pydefi._utils import decode_address
 from pydefi.aggregator.base import AggregatorQuote
 from pydefi.aggregator.jupiter import _JUPITER_API_BASE, _JUPITER_SWAP_V2_BASE, Jupiter, JupiterSwapV2
-from pydefi.amm.base import BaseSolanaAMM
 from pydefi.amm.raydium import _RAYDIUM_API_BASE, Raydium
 from pydefi.exceptions import AggregatorError, InsufficientLiquidityError
 from pydefi.types import Address, ChainId, SwapRoute, Token, TokenAmount
@@ -45,28 +45,21 @@ class TestChainIdSolana:
 
 
 # ---------------------------------------------------------------------------
-# BaseSolanaAMM abstract interface
+# BaseSolanaAMM was removed — concrete AMMs are duck-typed
 # ---------------------------------------------------------------------------
 
 
-class TestBaseSolanaAMM:
-    def test_cannot_instantiate_abstract(self):
-        with pytest.raises(TypeError):
-            BaseSolanaAMM()  # type: ignore[abstract]
-
+class TestSlippage:
     def test_apply_slippage(self):
-        # Test via a concrete subclass
-        raydium = Raydium()
-        assert raydium._apply_slippage(1_000_000, 50) == 995_000
-        assert raydium._apply_slippage(1_000_000, 0) == 1_000_000
-        assert raydium._apply_slippage(1_000_000, 10_000) == 0
+        assert apply_slippage(1_000_000, 50) == 995_000
+        assert apply_slippage(1_000_000, 0) == 1_000_000
+        assert apply_slippage(1_000_000, 10_000) == 0
 
     def test_apply_slippage_invalid(self):
-        raydium = Raydium()
         with pytest.raises(ValueError):
-            raydium._apply_slippage(1_000_000, -1)
+            apply_slippage(1_000_000, -1)
         with pytest.raises(ValueError):
-            raydium._apply_slippage(1_000_000, 10_001)
+            apply_slippage(1_000_000, 10_001)
 
 
 # ---------------------------------------------------------------------------
@@ -87,9 +80,9 @@ class TestRaydium:
         raydium = Raydium(api_url="https://my-raydium.example.com")
         assert raydium.api_url == "https://my-raydium.example.com"
 
-    def test_inherits_base_solana_amm(self):
+    def test_protocol_name_is_raydium(self):
         raydium = Raydium()
-        assert isinstance(raydium, BaseSolanaAMM)
+        assert raydium.protocol_name == "Raydium"
 
     @pytest.mark.asyncio
     async def test_get_quote_success(self):
@@ -316,10 +309,9 @@ class TestJupiter:
         headers = jupiter._headers()
         assert headers["Authorization"] == "Bearer test-key-123"
 
-    def test_inherits_base_aggregator(self):
-        from pydefi.aggregator.base import BaseAggregator
-
-        assert isinstance(Jupiter(), BaseAggregator)
+    def test_jupiter_is_callable(self):
+        j = Jupiter()
+        assert j.chain_id == ChainId.SOLANA
 
     @pytest.mark.asyncio
     async def test_get_quote_success(self):
@@ -547,10 +539,9 @@ class TestJupiterSwapV2:
         headers = JupiterSwapV2(api_key="my-api-key")._headers()
         assert headers["x-api-key"] == "my-api-key"
 
-    def test_inherits_base_aggregator(self):
-        from pydefi.aggregator.base import BaseAggregator
-
-        assert isinstance(JupiterSwapV2(), BaseAggregator)
+    def test_jupiter_swap_v2_is_callable(self):
+        j = JupiterSwapV2()
+        assert j.chain_id == ChainId.SOLANA
 
     @pytest.mark.asyncio
     async def test_get_order_without_taker(self):
