+++ tests/test_types.py
@@ -4,11 +4,11 @@ from decimal import Decimal
 
 import pytest
 
+from pydefi.pathfinder.dag import RouteDAG
 from pydefi.types import (
     Address,
     BridgeQuote,
     ChainId,
-    RouteDAG,
     RouteSplit,
     RouteSwap,
     SwapRoute,
